+++ src/android/com/synconset/cordovahttp/CordovaHttpDelete.java
@@ -1,40 +1,41 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
+package com.synconset.cordovahttp;
 
 import java.net.UnknownHostException;
-import java.util.Map;
-
-import org.apache.cordova.CallbackContext;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import javax.net.ssl.SSLHandshakeException;
 
-import android.util.Log;
+import org.apache.cordova.CallbackContext;
+
+import org.json.JSONException;
+import org.json.JSONObject;
 
 import com.github.kevinsawicki.http.HttpRequest;
 import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
- 
-public class CordovaHttpPost extends CordovaHttp implements Runnable {
-    public CordovaHttpPost(String urlString, Map<?, ?> params, Map<String, String> headers, CallbackContext callbackContext) {
-        super(urlString, params, headers, callbackContext);
+
+class CordovaHttpDelete extends CordovaHttp implements Runnable {
+    public CordovaHttpDelete(String urlString, JSONObject data, JSONObject headers, CallbackContext callbackContext) {
+        super(urlString, data, headers, callbackContext);
     }
-    
+
     @Override
     public void run() {
         try {
-            HttpRequest request = HttpRequest.post(this.getUrlString());
+            HttpRequest request = HttpRequest.delete(this.getUrlString(), this.getParamsMap(), false);
+
             this.setupSecurity(request);
             request.acceptCharset(CHARSET);
-            request.headers(this.getHeaders());
-            request.form(this.getParams());
+            request.headers(this.getHeadersMap());
+
             int code = request.code();
             String body = request.body(CHARSET);
             JSONObject response = new JSONObject();
+
             this.addResponseHeaders(request, response);
             response.put("status", code);
+
             if (code >= 200 && code < 300) {
                 response.put("data", body);
                 this.getCallbackContext().success(response);
@@ -44,7 +45,7 @@ public class CordovaHttpPost extends CordovaHttp implements Runnable {
             }
         } catch (JSONException e) {
             this.respondWithError("There was an error generating the response");
-        }  catch (HttpRequestException e) {
+        } catch (HttpRequestException e) {
             if (e.getCause() instanceof UnknownHostException) {
                 this.respondWithError(0, "The host could not be resolved");
             } else if (e.getCause() instanceof SSLHandshakeException) {