+++ src/android/com/synconset/cordovahttp/CordovaHttpDownload.java
@@ -1,9 +1,7 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
-
-import android.util.Log;
+package com.synconset.cordovahttp;
 
 import com.github.kevinsawicki.http.HttpRequest;
 import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
@@ -12,7 +10,6 @@ import java.io.File;
 import java.net.UnknownHostException;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.Map;
 
 import javax.net.ssl.SSLHandshakeException;
 
@@ -22,23 +19,23 @@ import org.apache.cordova.file.FileUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class CordovaHttpDownload extends CordovaHttp implements Runnable {
+class CordovaHttpDownload extends CordovaHttp implements Runnable {
     private String filePath;
-    
-    public CordovaHttpDownload(String urlString, Map<?, ?> params, Map<String, String> headers, CallbackContext callbackContext, String filePath) {
+
+    public CordovaHttpDownload(String urlString, JSONObject params, JSONObject headers, CallbackContext callbackContext, String filePath) {
         super(urlString, params, headers, callbackContext);
         this.filePath = filePath;
     }
-    
+
     @Override
     public void run() {
         try {
-            HttpRequest request = HttpRequest.get(this.getUrlString(), this.getParams(), true);
+            HttpRequest request = HttpRequest.get(this.getUrlString(), this.getParamsMap(), true);
             this.setupSecurity(request);
             request.acceptCharset(CHARSET);
-            request.headers(this.getHeaders());
+            request.headers(this.getHeadersMap());
             int code = request.code();
-            
+
             JSONObject response = new JSONObject();
             this.addResponseHeaders(request, response);
             response.put("status", code);