+++ src/android/com/synconset/cordovahttp/CordovaHttpPost.java
@@ -1,57 +1,57 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
+package com.synconset.cordovahttp;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.net.UnknownHostException;
-import java.util.Map;
-
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.HostnameVerifier;
-
-import javax.net.ssl.SSLHandshakeException;
 
 import org.apache.cordova.CallbackContext;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import android.util.Log;
+import javax.net.ssl.SSLHandshakeException;
 
 import com.github.kevinsawicki.http.HttpRequest;
 import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
- 
-public class CordovaHttpHead extends CordovaHttp implements Runnable {
-    public CordovaHttpHead(String urlString, Map<?, ?> params, Map<String, String> headers, CallbackContext callbackContext) {
-        super(urlString, params, headers, callbackContext);
+
+class CordovaHttpPost extends CordovaHttp implements Runnable {
+    public CordovaHttpPost(String urlString, JSONObject params, String serializerName, JSONObject headers, CallbackContext callbackContext) {
+        super(urlString, params, serializerName, headers, callbackContext);
     }
-    
+
     @Override
     public void run() {
         try {
-            HttpRequest request = HttpRequest.head(this.getUrlString(), this.getParams(), true);
+            HttpRequest request = HttpRequest.post(this.getUrlString());
+
             this.setupSecurity(request);
             request.acceptCharset(CHARSET);
-            request.headers(this.getHeaders());
+            request.headers(this.getHeadersMap());
+
+            if (new String("json").equals(this.getSerializerName())) {
+                request.contentType(request.CONTENT_TYPE_JSON, request.CHARSET_UTF8);
+                request.send(this.getParamsObject().toString());
+            } else {
+                request.form(this.getParamsMap());
+            }
+
             int code = request.code();
+            String body = request.body(CHARSET);
             JSONObject response = new JSONObject();
+
             this.addResponseHeaders(request, response);
             response.put("status", code);
+
             if (code >= 200 && code < 300) {
-                // no 'body' to return for HEAD request
+                response.put("data", body);
                 this.getCallbackContext().success(response);
             } else {
-                String body = request.body(CHARSET);
                 response.put("error", body);
                 this.getCallbackContext().error(response);
             }
         } catch (JSONException e) {
             this.respondWithError("There was an error generating the response");
-        } catch (HttpRequestException e) {
+        }  catch (HttpRequestException e) {
             if (e.getCause() instanceof UnknownHostException) {
                 this.respondWithError(0, "The host could not be resolved");
             } else if (e.getCause() instanceof SSLHandshakeException) {